Server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Mukarramboy Umarov",email = "mukarramboyumarov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-filter (>=25.1,<26.0)",
    "drf-yasg (>=1.21.11,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
